More work on EquityPortfolio, but need to agree on path forward first
Question 1: When calculating dividend dates, all dates are in a single list. Would it make more sense to put an extra layer to differentiate dictionaries from different bonds?
Question 2: Is it worth using numpy arrays so that we can do vector subtractions or should we get into a habit of looping through them?